Ignore variants during JS codegen if they are configured with `as_direct_union_of_resource_classes` (#16)
Right now we _mostly_ ignore them, except when the resource classes include static methods. In that case code is meant to be generated, but it was not implemented and so would crash instead. For now we'll just completely ignore it and not try to generate anything.

This change allows us to add static methods to resources which are part of variants.